Conflicts:
	main/external/Mono.Cecil
	main/src/core/MonoDevelop.Startup/MonoDevelop.Startup.csproj

See http://stackoverflow.com/questions/826715/#comment16042007_836019
for why the main/external/Mono.Cecil submodule was a PITA

git checkout master main/external/Mono.Cecil

+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.TypeSystem/TypeSystemService.cs
@@ -145,15 +145,26 @@ namespace MonoDevelop.Ide.TypeSystem
 	
 	public static class TypeSystemService
 	{
-		const string CurrentVersion = "1.0";
+		const string CurrentVersion = "1.0.1";
 		static List<TypeSystemParserNode> parsers;
-		public static readonly HashSet<string> FilesSkippedInParseThread = new HashSet<string> ();
+		static string[] filesSkippedInParseThread = new string[0];
 		static IEnumerable<TypeSystemParserNode> Parsers {
 			get {
 				return parsers;
 			}
 		}
-		
+
+		public static void RemoveSkippedfile (FilePath fileName)
+		{
+			filesSkippedInParseThread = filesSkippedInParseThread.Where (f => f != fileName).ToArray ();
+		}
+		public static void AddSkippedFile (FilePath fileName)
+		{
+			if (filesSkippedInParseThread.Any (f => f == fileName))
+				return;
+			filesSkippedInParseThread = filesSkippedInParseThread.Concat (new string[] { fileName }).ToArray ();
+		}
+
 		static TypeSystemService ()
 		{
 			parsers = new List<TypeSystemParserNode> ();
@@ -258,8 +269,12 @@ namespace MonoDevelop.Ide.TypeSystem
 						wrapper = null;
 					}
 					if (wrapper != null && (result.Flags & ParsedDocumentFlags.NonSerializable) != ParsedDocumentFlags.NonSerializable) {
+						var oldFile = wrapper.Content.GetFile (fileName);
 						wrapper.UpdateContent (c => c.UpdateProjectContent (c.GetFile (fileName), result.ParsedFile));
 						UpdateParsedDocument (wrapper, result);
+						if (oldFile != null)
+							wrapper.InformFileRemoved (new ParsedFileEventArgs (oldFile));
+						wrapper.InformFileAdded (new ParsedFileEventArgs (result.ParsedFile));
 					}
 
 					// The parsed file could be included in other projects as well, therefore
@@ -271,6 +286,8 @@ namespace MonoDevelop.Ide.TypeSystem
 						var file = cnt.Value.Content.GetFile (fileName);
 						if (file != null) {
 							cnt.Value.UpdateContent (c => c.UpdateProjectContent (file, result.ParsedFile));
+							cnt.Value.InformFileRemoved (new ParsedFileEventArgs (file));
+							cnt.Value.InformFileAdded (new ParsedFileEventArgs (result.ParsedFile));
 						}
 					}
 				}
@@ -307,8 +324,12 @@ namespace MonoDevelop.Ide.TypeSystem
 				var result = parser.Parse (true, fileName, content);
 				lock (projectWrapperUpdateLock) {
 					if (wrapper != null && (result.Flags & ParsedDocumentFlags.NonSerializable) != ParsedDocumentFlags.NonSerializable) {
+						var oldFile = wrapper.Content.GetFile (fileName);
 						wrapper.UpdateContent (c => c.UpdateProjectContent (c.GetFile (fileName), result.ParsedFile));
 						UpdateParsedDocument (wrapper, result);
+						if (oldFile != null)
+							wrapper.InformFileRemoved (new ParsedFileEventArgs (oldFile));
+						wrapper.InformFileAdded (new ParsedFileEventArgs (result.ParsedFile));
 					}
 				}
 				return result;
@@ -636,9 +657,6 @@ namespace MonoDevelop.Ide.TypeSystem
 				get {
 					return assembly.Location;
 				}
-				set {
-					assembly.Location = value;
-				}
 			}
 
 			public IEnumerable<IUnresolvedAttribute> AssemblyAttributes {
@@ -752,7 +770,25 @@ namespace MonoDevelop.Ide.TypeSystem
 					WasChanged = true;
 				}
 			}
-			
+
+			public void InformFileRemoved (ParsedFileEventArgs e)
+			{
+				var handler = FileRemoved;
+				if (handler != null)
+					handler (this, e);
+			}
+
+			public void InformFileAdded (ParsedFileEventArgs e)
+			{
+				var handler = FileAdded;
+				if (handler != null)
+					handler (this, e);
+			}
+
+			public EventHandler<ParsedFileEventArgs> FileAdded;
+			public EventHandler<ParsedFileEventArgs> FileRemoved;
+
+
 			public bool WasChanged = false;
 			
 			[NonSerialized]
@@ -815,16 +851,19 @@ namespace MonoDevelop.Ide.TypeSystem
 					this.wrapper = wrapper;
 					contextTask = Task.Factory.StartNew (delegate {
 	
-						if (projectCache.ContainsKey (this.wrapper.Project.FileName))
-							return projectCache [this.wrapper.Project.FileName];
+						IProjectContent content;
+						if (projectCache.TryGetValue (this.wrapper.Project.FileName, out content)) {
+							if (content != null)
+								return content;
+						}
 	
 						var context = LoadProjectCache (this.wrapper.Project);
 						if (context != null) {
-							return context.SetAssemblyName (this.wrapper.Project.Name);
+							return context.SetAssemblyName (this.wrapper.Project.Name) ?? context;
 						}
 
 						context = new CSharpProjectContent ();
-						context.Location = wrapper.Project.FileName;
+						context = context.SetLocation (this.wrapper.Project.FileName);
 						context = context.SetAssemblyName (this.wrapper.Project.Name);
 						QueueParseJob (this.wrapper);
 						return context;
@@ -866,9 +905,6 @@ namespace MonoDevelop.Ide.TypeSystem
 					get {
 						return Content.Location;
 					}
-					set {
-						Content.Location = value;
-					}
 				}
 
 				IEnumerable<IUnresolvedAttribute> IUnresolvedAssembly.AssemblyAttributes {
@@ -891,7 +927,10 @@ namespace MonoDevelop.Ide.TypeSystem
 				#endregion
 
 				#region IProjectContent implementation
-				IParsedFile IProjectContent.GetFile (string fileName)
+
+				string IProjectContent.ProjectFileName { get { return Content.ProjectFileName; } }
+
+				IUnresolvedFile IProjectContent.GetFile (string fileName)
 				{
 					return Content.GetFile (fileName);
 				}
@@ -911,7 +950,17 @@ namespace MonoDevelop.Ide.TypeSystem
 					return Content.SetAssemblyName (newAssemblyName);
 				}
 
+				IProjectContent IProjectContent.SetLocation (string newLocation)
+				{
+					return Content.SetLocation (newLocation);
+				}
+
 				IProjectContent IProjectContent.AddAssemblyReferences (IEnumerable<IAssemblyReference> references)
+				{
+					return Content.AddAssemblyReferences (references);
+				}
+
+				IProjectContent IProjectContent.AddAssemblyReferences (params IAssemblyReference[] references)
 				{
 					return Content.AddAssemblyReferences (references);
 				}
@@ -921,17 +970,32 @@ namespace MonoDevelop.Ide.TypeSystem
 					return Content.RemoveAssemblyReferences (references);
 				}
 
-				IProjectContent IProjectContent.UpdateProjectContent (IParsedFile oldFile, IParsedFile newFile)
+				IProjectContent IProjectContent.RemoveAssemblyReferences (params IAssemblyReference[] references)
+				{
+					return Content.RemoveAssemblyReferences (references);
+				}
+
+				IProjectContent IProjectContent.UpdateProjectContent (IUnresolvedFile oldFile, IUnresolvedFile newFile)
 				{
 					return Content.UpdateProjectContent (oldFile, newFile);
 				}
 
-				public IProjectContent UpdateProjectContent (IEnumerable<IParsedFile> oldFiles, IEnumerable<IParsedFile> newFiles)
+				public IProjectContent UpdateProjectContent (IEnumerable<IUnresolvedFile> oldFiles, IEnumerable<IUnresolvedFile> newFiles)
 				{
 					return Content.UpdateProjectContent (oldFiles, newFiles);
 				}
 
-				IEnumerable<IParsedFile> IProjectContent.Files {
+				public IProjectContent AddOrUpdateFiles (IEnumerable<IUnresolvedFile> newFiles)
+				{
+					return Content.AddOrUpdateFiles (newFiles);
+				}
+
+				public IProjectContent AddOrUpdateFiles (params IUnresolvedFile[] newFiles)
+				{
+					return Content.AddOrUpdateFiles (newFiles);
+				}
+
+				IEnumerable<IUnresolvedFile> IProjectContent.Files {
 					get {
 						return Content.Files;
 					}
@@ -941,6 +1005,21 @@ namespace MonoDevelop.Ide.TypeSystem
 					get {
 						return Content.AssemblyReferences;
 					}
+				}
+
+				IProjectContent IProjectContent.SetProjectFileName (string newProjectFileName)
+				{
+					return Content.SetProjectFileName (newProjectFileName);
+				}
+
+				IProjectContent IProjectContent.RemoveFiles (IEnumerable<string> fileNames)
+				{
+					return Content.RemoveFiles (fileNames);
+				}
+
+				IProjectContent IProjectContent.RemoveFiles (params string[] fileNames)
+				{
+					return Content.RemoveFiles (fileNames);
 				}
 				#endregion
 
@@ -1046,8 +1125,9 @@ namespace MonoDevelop.Ide.TypeSystem
 					project.FileRemovedFromProject += OnFileRemoved;
 					project.FileRenamedInProject += OnFileRenamed;
 					project.Modified += OnProjectModified;
+					var files = project.Files.ToArray ();
 					Task.Factory.StartNew (delegate {
-						CheckModifiedFiles (project, wrapper);
+						CheckModifiedFiles (project, files, wrapper);
 					});
 
 					return wrapper;
@@ -1088,7 +1168,12 @@ namespace MonoDevelop.Ide.TypeSystem
 			foreach (ProjectFileEventInfo fargs in args) {
 				var wrapper = projectContents [project];
 				var fileName = fargs.ProjectFile.Name;
+				var file = wrapper.Content.GetFile (fileName);
+				if (file == null)
+					continue;
 				wrapper.UpdateContent (c => c.UpdateProjectContent (c.GetFile (fileName), null));
+				wrapper.InformFileRemoved (new ParsedFileEventArgs (file));
+
 				var tags = wrapper.GetExtensionObject <ProjectCommentTags> ();
 				if (tags != null)
 					tags.RemoveFile (wrapper.Project, fileName);
@@ -1099,8 +1184,14 @@ namespace MonoDevelop.Ide.TypeSystem
 		{
 			var project = (Project)sender;
 			foreach (ProjectFileRenamedEventInfo fargs in args) {
-				projectContents [project].UpdateContent (c => c.UpdateProjectContent (c.GetFile (fargs.OldName), null));
-				QueueParseJob (projectContents [project], new [] { fargs.ProjectFile });
+				var content = projectContents [project];
+				var file = content.Content.GetFile (fargs.OldName);
+				if (file == null)
+					continue;
+				content.UpdateContent (c => c.UpdateProjectContent (c.GetFile (fargs.OldName), null));
+				content.InformFileRemoved (new ParsedFileEventArgs (file));
+
+				QueueParseJob (content, new [] { fargs.ProjectFile });
 			}
 		}
 		
@@ -1147,18 +1238,20 @@ namespace MonoDevelop.Ide.TypeSystem
 		{
 			if (DecLoadCount (project) != 0)
 				return;
-			if (referenceCounter.ContainsKey (project) && --referenceCounter [project] <= 0) {
-				project.FileAddedToProject -= OnFileAdded;
-				project.FileRemovedFromProject -= OnFileRemoved;
-				project.FileRenamedInProject -= OnFileRenamed;
-				project.Modified -= OnProjectModified;
-				
-				var wrapper = projectContents [project];
-				projectContents.Remove (project);
-				referenceCounter.Remove (project);
-				StoreProjectCache (project, wrapper);
-				
-				OnProjectUnloaded (new ProjectUnloadEventArgs (project, wrapper));
+			lock (projectWrapperUpdateLock) {
+				if (referenceCounter.ContainsKey (project) && --referenceCounter [project] <= 0) {
+					project.FileAddedToProject -= OnFileAdded;
+					project.FileRemovedFromProject -= OnFileRemoved;
+					project.FileRenamedInProject -= OnFileRenamed;
+					project.Modified -= OnProjectModified;
+					
+					var wrapper = projectContents [project];
+					projectContents.Remove (project);
+					referenceCounter.Remove (project);
+
+					StoreProjectCache (project, wrapper);
+					OnProjectUnloaded (new ProjectUnloadEventArgs (project, wrapper));
+				}
 			}
 		}
 		
@@ -1254,6 +1347,7 @@ namespace MonoDevelop.Ide.TypeSystem
 					return result;
 				
 				var name = AssemblyNameReference.Parse (fullName);
+				AssemblyDefinition bestFit = null;
 				// need to handle different file extension casings. Some dlls from windows tend to end with .Dll or .DLL rather than '.dll'
 				foreach (string file in Directory.GetFiles (lookupPath, name.Name + ".*")) {
 					string ext = Path.GetExtension (file);
@@ -1262,6 +1356,11 @@ namespace MonoDevelop.Ide.TypeSystem
 					ext = ext.ToUpper ();
 					if (ext == ".DLL" || ext == ".EXE") {
 						result = ReadAssembly (file);
+						if (result.FullName != fullName) {
+							bestFit = result;
+							result = null;
+							continue;
+						}
 						break;
 					}
 				}
@@ -1275,6 +1374,8 @@ namespace MonoDevelop.Ide.TypeSystem
 						result = ReadAssembly (location);
 					}
 				}
+				if (result == null)
+					result = bestFit;
 				if (result != null)
 					cache [fullName] = result;
 				return result;
@@ -1413,9 +1514,6 @@ namespace MonoDevelop.Ide.TypeSystem
 				get {
 					return Ctx.Location;
 				}
-				set {
-					Ctx.Location = value;
-				}
 			}
 
 			IEnumerable<IUnresolvedAttribute> IUnresolvedAssembly.AssemblyAttributes {
@@ -1483,9 +1581,9 @@ namespace MonoDevelop.Ide.TypeSystem
 					if (File.Exists (assemblyPath)) {
 						var deserializedAssembly = DeserializeObject <IUnresolvedAssembly> (assemblyPath);
 						if (deserializedAssembly != null) {
-							var provider = deserializedAssembly as IDocumentationProviderContainer;
+						/*	var provider = deserializedAssembly as IDocumentationProviderContainer;
 							if (provider != null)
-								provider.DocumentationProvider = new CombinedDocumentationProvider (fileName);
+								provider.DocumentationProvider = new CombinedDocumentationProvider (fileName);*/
 							return deserializedAssembly;
 						}
 					}
@@ -1501,7 +1599,6 @@ namespace MonoDevelop.Ide.TypeSystem
 					loader.IncludeInternalMembers = true;
 					loader.DocumentationProvider = new CombinedDocumentationProvider (fileName);
 					assembly = loader.LoadAssembly (asm);
-					assembly.Location = fileName;
 				} catch (Exception e) {
 					LoggingService.LogError ("Can't convert assembly: " + fileName, e);
 					return null;
@@ -1625,8 +1722,6 @@ namespace MonoDevelop.Ide.TypeSystem
 				throw new ArgumentNullException ("project");
 			ProjectContentWrapper content;
 			if (projectContents.TryGetValue (project, out content)) {
-				if (content.Content != null)
-					content.Content.Location = project.FileName;
 				return content;
 			}
 			return new ProjectContentWrapper (project);
@@ -1694,24 +1789,26 @@ namespace MonoDevelop.Ide.TypeSystem
 			{
 				TypeSystemParserNode node = null;
 				ITypeSystemParser parser = null;
-				lock (FilesSkippedInParseThread) {
-					foreach (var file in (FileList ?? Context.Project.Files)) {
-						if (!string.Equals (file.BuildAction, "compile", StringComparison.OrdinalIgnoreCase)) 
-							continue;
-						var fileName = file.FilePath;
-						if (FilesSkippedInParseThread.Contains (fileName))
-							continue;
-						if (node == null || !node.CanParse (fileName)) {
-							node = TypeSystemService.GetTypeSystemParserNode (DesktopService.GetMimeTypeForUri (fileName));
-							parser = node != null ? node.Parser : null;
-						}
-						if (parser == null)
-							continue;
-						using (var stream = new System.IO.StreamReader (fileName)) {
-							var parsedDocument = parser.Parse (false, fileName, stream, Context.Project);
-							UpdateParsedDocument (Context, parsedDocument);
-							Context.UpdateContent (c => c.UpdateProjectContent (c.GetFile (fileName), parsedDocument.ParsedFile));
-						}
+				foreach (var file in (FileList ?? Context.Project.Files)) {
+					if (!string.Equals (file.BuildAction, "compile", StringComparison.OrdinalIgnoreCase)) 
+						continue;
+					var fileName = file.FilePath;
+					if (filesSkippedInParseThread.Any (f => f == fileName))
+						continue;
+					if (node == null || !node.CanParse (fileName)) {
+						node = TypeSystemService.GetTypeSystemParserNode (DesktopService.GetMimeTypeForUri (fileName));
+						parser = node != null ? node.Parser : null;
+					}
+					if (parser == null)
+						continue;
+					using (var stream = new System.IO.StreamReader (fileName)) {
+						var parsedDocument = parser.Parse (false, fileName, stream, Context.Project);
+						UpdateParsedDocument (Context, parsedDocument);
+						var oldFile = Context.Content.GetFile (fileName);
+						Context.UpdateContent (c => c.UpdateProjectContent (c.GetFile (fileName), parsedDocument.ParsedFile));
+						if (oldFile != null)
+							Context.InformFileRemoved (new ParsedFileEventArgs (oldFile));
+						Context.InformFileAdded (new ParsedFileEventArgs (parsedDocument.ParsedFile));
 					}
 				}
 			}
@@ -1859,7 +1956,7 @@ namespace MonoDevelop.Ide.TypeSystem
 			}
 		}
 
-		static bool IsFileModified (ProjectFile file, IParsedFile parsedFile)
+		static bool IsFileModified (ProjectFile file, IUnresolvedFile parsedFile)
 		{
 			if (parsedFile == null)
 				return true;
@@ -1870,36 +1967,44 @@ namespace MonoDevelop.Ide.TypeSystem
 			}
 		}
 
-		static void CheckModifiedFiles (Project project, ProjectContentWrapper content)
+		static void CheckModifiedFiles (Project project, ProjectFile[] projectFiles, ProjectContentWrapper content)
 		{
-			List<ProjectFile> modifiedFiles = null;
-			foreach (var file in project.Files) {
-				if (!string.Equals (file.BuildAction, "compile", StringComparison.OrdinalIgnoreCase)) 
-					continue;
-				var fileName = file.Name;
-				// if the file is already inside the content a parser exists for it, if not check if it can be parsed.
-				var oldFile = content.Content.GetFile (fileName);
-				if (oldFile == null) {
-					var parser = TypeSystemService.GetParser (DesktopService.GetMimeTypeForUri (fileName));
-					if (parser == null)
-						continue;
+			try {
+				lock (projectWrapperUpdateLock) {
+					List<ProjectFile> modifiedFiles = null;
+					foreach (var file in projectFiles) {
+						if (file.BuildAction == null || !string.Equals (file.BuildAction, "compile", StringComparison.OrdinalIgnoreCase)) 
+							continue;
+						var fileName = file.Name;
+						// if the file is already inside the content a parser exists for it, if not check if it can be parsed.
+						var oldFile = content.Content.GetFile (fileName);
+						if (oldFile == null) {
+							var parser = TypeSystemService.GetParser (DesktopService.GetMimeTypeForUri (fileName));
+							if (parser == null)
+								continue;
+						}
+						if (!IsFileModified (file, oldFile))
+							continue;
+						if (modifiedFiles == null)
+							modifiedFiles = new List<ProjectFile> ();
+						modifiedFiles.Add (file);
+					}
+					
+					// check if file needs to be removed from project content 
+					foreach (var file in content.Content.Files) {
+						if (project.GetProjectFile (file.FileName) == null) {
+							content.UpdateContent (c => c.UpdateProjectContent (file, null));
+							content.InformFileRemoved (new ParsedFileEventArgs (file));
+						}
+					}
+					
+					if (modifiedFiles == null)
+						return;
+					QueueParseJob (content, modifiedFiles);
 				}
-				if (!IsFileModified (file, oldFile))
-					continue;
-				if (modifiedFiles == null)
-					modifiedFiles = new List<ProjectFile> ();
-				modifiedFiles.Add (file);
+			} catch (Exception e) {
+				LoggingService.LogError ("Exception in check modified files.", e);
 			}
-			
-			// check if file needs to be removed from project content 
-			foreach (var file in content.Content.Files) {
-				if (project.GetProjectFile (file.FileName) == null)
-					content.UpdateContent (c => c.UpdateProjectContent (file, null));
-			}
-			
-			if (modifiedFiles == null)
-				return;
-			QueueParseJob (content, modifiedFiles);
 		}
 
 		static void CheckModifiedFile (AssemblyContext context)
@@ -1934,7 +2039,8 @@ namespace MonoDevelop.Ide.TypeSystem
 			
 			while (list.Count > 0) {
 				var readydb = list.Dequeue ();
-				CheckModifiedFiles (readydb.Key, readydb.Value);
+				var files = readydb.Key.Files.ToArray ();
+				CheckModifiedFiles (readydb.Key, files, readydb.Value);
 			}
 			
 			Queue<KeyValuePair<string, AssemblyContext>> assemblyList;
